公演期間クラスの追加、および公演開始日公演終了日クラスの削除を行いました。
